# Insight-Performance-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Organisational Performance Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notification-configs:
    get:
      operationId: listNotificationConfigs
      tags: [notifications]
      summary: List notification configurations
      parameters:
        - name: cycleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notification configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationConfig"
    post:
      operationId: createNotificationConfig
      tags: [notifications]
      summary: Create a notification configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationConfigInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationConfig"
  /notification-configs/{id}:
    patch:
      operationId: updateNotificationConfig
      tags: [notifications]
      summary: Update a notification configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationConfig"
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [reminder, escalation, info, warning]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
    NotificationConfig:
      type: object
      properties:
        id:
          type: integer
        cycleId:
          type: integer
        eventType:
          type: string
        daysBefore:
          type: integer
        isEmail:
          type: boolean
        isInApp:
          type: boolean
        isActive:
          type: boolean
      required:
        - id
        - cycleId
        - eventType
        - daysBefore
        - isEmail
        - isInApp
        - isActive
    CreateNotificationConfigInput:
      type: object
      properties:
        cycleId:
          type: integer
        eventType:
          type: string
        daysBefore:
          type: integer
          default: 7
        isEmail:
          type: boolean
          default: false
        isInApp:
          type: boolean
          default: true
        isActive:
          type: boolean
          default: true
      required:
        - cycleId
        - eventType
    UpdateNotificationConfigInput:
      type: object
      properties:
        daysBefore:
          type: integer
        isEmail:
          type: boolean
        isInApp:
          type: boolean
        isActive:
          type: boolean
...